Moved interpreter to ./src/lib
