+++ src/lib/utilities.c
